colocate traits of Amount alongside struct
